[#11] Consolidate Redis type management into a single module
(Related to the previous 2 commits)

Improve modularity and maintainability by:
- Merging Redis type and relation management into RelationsManagement
- Removing redundant files and modules
- Moving class-level instance variables to the top of the Horreum class
- Extending Settings and RelationsManagement in ClassMethods
- Calling setup_relations_management to initialize the feature

These changes simplify the codebase and provide a clear separation between the feature's interface and its internal implementation.
