+++ lib/familia/horreum.rb
@@ -86,7 +86,7 @@ module Familia
       elsif kwargs.any?
         initialize_with_keyword_args(**kwargs)
       else
-        Familia.debug "[Horreum] #{self.class} initialized with no arguments"
+        Familia.ld "[Horreum] #{self.class} initialized with no arguments"
         # If there are no arguments, we need to set the default values
         # for the fields. This is done in the order they were defined.
         # self.class.fields.each do |field|
